Refactor Deck structure and update JDK version
The primary change in this commit is the refactoring of the 'Deck' class structure for the Yu-Gi-Oh card game. The 'Deck' has been divided into 'AbstractDeck', 'ExtraDeck', and 'MainDeck' subclasses to better encapsulate respective functionalities. Test classes for these subclasses have also been added. The JDK version in the project settings was downgraded to 17 from the default 22. Also, the 'Card' class was moved to a new 'cards' package.